requirements.txt
Django==4.2.0
pymongo==4.6.0
pandas==2.0.0
scikit-learn==1.3.0
numpy==1.24.0
dnspython==2.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
